# openbank-delegation-service/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: OpenBank Delegation Service API
  description: |
    Customer-to-party delegated access (ADR-0232): granular, SCA-bound grants over
    products (accounts, savings goals, cards) and single objects (payment, statement,
    document). Lifecycle: OFFERED → ACTIVE → SUSPENDED/REVOKED/EXPIRED/RENOUNCED/DECLINED.
  # 1.13.0 -> 1.14.0 (#9286): additive — customer-managed delegated account portfolios.
  #
  # 1.14.0 -> 1.18.0 (#9522, this branch): additive — organization grant authority. NOT the next
  # integer: 1.15.0 is held by the open #9215, 1.16.0 by #9481/#9484 and 1.17.0 by #9237. Read off
  # live origin/main together with the open PRs, because a taken version is textually identical to
  # a free one and no merge can catch the collision.
  version: 1.18.0
servers:
  - url: http://localhost:8126
...
paths:
...
  /api/v1/delegation-role-presets/{id}:
    parameters:
      - $ref: '#/components/parameters/DelegationId'
    put:
      tags: [Delegation role presets]
      summary: Replace a delegation role preset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DelegationRolePresetRequest' }
      responses:
        '200': { description: Preset updated }
        '404': { description: Preset not found }
    delete:
      tags: [Delegation role presets]
      summary: Delete a delegation role preset
      responses:
        '204': { description: Preset deleted }
        '404': { description: Preset not found }
  /api/v1/delegations:
    post:
      tags: [Delegations]
      summary: Offer a delegation grant (OFFERED); idempotent via X-Request-ID (ADR-0292)
      parameters:
        - name: X-Request-ID
          in: header
          schema: { type: string }
        - $ref: '#/components/parameters/CustomerPartyId'
        - $ref: '#/components/parameters/CustomerActorPartyId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OfferDelegationRequest' }
      responses:
        '201':
          description: Grant offered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DelegationResponse' }
        '400':
          description: |
            Validation or SCA failure. Machine-readable `code`:

            * `CUMULATIVE_LIMIT_UNSUPPORTED` — dailyLimit/monthlyLimit on a grant with no
              money-moving capability, where nothing would ever reserve against them.
            * `SPEND_WITHOUT_CEILING` — ACCOUNT_INITIATE_PAYMENT with neither dailyLimit nor
              monthlyLimit (ADR-0249 D5). Unlimited access to someone else's account is not
              something this platform grants by omission.
            * `APPROVAL_POLICY_UNSUPPORTED` — approvalPolicy other than SOLO; nothing counts approvals.
            * `EXPOSURE_UNSUPPORTED` — exposure is non-null. It remains readable on historical grants
              for audit, but cannot authorize a raw object download until transformed-artifact delivery
              can prove redaction, watermark and view-count enforcement.

            All are refused before the SCA challenge is spent, so the customer keeps their ceremony.
        '403': { description: The authenticated party is not the grantor, or the actor has no active authority for it }
        '422': { description: Eligibility gate, or the grantor does not own the resource (ADR-0232 D5, threat model T1) }
        '503': { description: Grantor authority could not be verified; retry is safe because SCA was not consumed }
...
  /api/v1/delegations/approvals/{approvalId}:
    get:
      tags: [Delegation lifecycle approvals]
      summary: Get immutable lifecycle approval evidence
      parameters:
        - $ref: '#/components/parameters/ApprovalId'
      responses:
        '200':
          description: Immutable proposal, decision and execution evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DelegationLifecycleApprovalResponse' }
        '403': { description: A human ROLE_OPERATOR or ROLE_ADMIN session is required }
        '404': { description: Approval not found }
...
  /api/v1/delegations/{id}/reservations:
    post:
      tags: [Delegations]
      summary: Reserve headroom under the grant's per-transaction, daily and monthly ceilings
      description: |
        ADR-0249 D3 — the authoritative cumulative-spend counter. The caller reserves BEFORE
        initiating the payment, confirms on a settled outcome and releases on failure. A RESERVED
        reservation already consumes headroom; that is the point, and it is why two concurrent
        payments cannot both pass a check that neither passes alone.

        A retry under the same `idempotencyKey` and immutable amount/currency/operation tuple
        returns the same reservation. A different tuple is refused. `operationType` defaults to
        `UNSPECIFIED`; `DOMESTIC_PAYMENT` opts into the durable state stream and remains disabled
        until its downstream binding consumer is ready.
      parameters:
        - $ref: '#/components/parameters/DelegationId'
        - $ref: '#/components/parameters/CustomerPartyId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveSpendRequest' }
      responses:
        '201':
          description: Headroom reserved (or an existing reservation replayed under the same key)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpendReservationResponse' }
        '404': { description: No such delegation, or the caller is not its grantee }
        '400': { description: Malformed request, including an idempotency key above its rail limit }
        '409':
          description: >-
            Refused by a ceiling, or the idempotency key names a different immutable tuple.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpendRefusalResponse' }
            application/problem+json:
              schema: { $ref: '#/components/schemas/SpendIdempotencyConflictResponse' }
        '503':
          description: Domestic reservation admission is disabled; no reservation was created.
  /api/v1/delegations/{id}/reservations/{rid}/confirm:
    post:
      tags: [Delegations]
      summary: Confirm a reservation — the money moved, the headroom stays consumed
      parameters:
        - $ref: '#/components/parameters/DelegationId'
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/CustomerPartyId'
      responses:
        '200':
          description: Confirmed (idempotent on an already-confirmed reservation)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpendReservationResponse' }
        '404': { description: No such reservation on this delegation }
        '409': { description: The reservation was released and cannot be confirmed }
  /api/v1/delegations/{id}/reservations/{rid}/release:
    post:
      tags: [Delegations]
      summary: Release a reservation — the payment did not happen, the headroom comes back
      description: >-
        Must be called on every failure branch. A leaked reservation silently shrinks the
        delegate's ceiling, which is why release is tested alongside confirm rather than after it.
      parameters:
        - $ref: '#/components/parameters/DelegationId'
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/CustomerPartyId'
      responses:
        '200':
          description: Released (idempotent on an already-released reservation)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpendReservationResponse' }
        '404': { description: No such reservation on this delegation }
        '409':
          description: >-
            The reservation was confirmed. Releasing settled money would re-open a ceiling the
            delegate has already spent through.
components:
  parameters:
    ApprovalId:
      name: approvalId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CustomerPartyId:
      name: X-Customer-Party-Id
      in: header
      required: false
      description: |
        The party the customer channel authenticated, stamped by customer-edge on every
        customer-scoped call (the fleet's IDOR-guard convention). When present, the service
        refuses any operation whose claimed party differs from it — a party id in a body or
        query string is a claim by the caller, this header is not.

        Absent on an operator/back-office call, which is gated by role and OPA instead.
      schema: { type: string, format: uuid }
    CustomerActorPartyId:
      name: X-Customer-Actor-Party-Id
      in: header
      required: false
      description: |
        The natural person authenticated by customer-edge. Required together with
        X-Customer-Party-Id on customer preview/offer calls. It equals the principal for an
        individual profile; for an organization profile it differs and the service verifies the
        actor's active PartyMandate before accepting the request. Never supplied by the app.
      schema: { type: string, format: uuid }
    DelegationId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ReservationId:
      name: rid
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    DelegationRolePresetRequest:
      type: object
      required: [name, resourceType, capabilities]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 500 }
        resourceType: { type: string, enum: [ACCOUNT, SAVINGS_GOAL, CARD, PAYMENT, STATEMENT, DOCUMENT] }
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
...
    MoneyDto:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string }
        currency: { type: string, minLength: 3, maxLength: 3 }
    ExposureDto:
      type: object
      properties:
        redactionRules: { type: array, items: { type: string } }
        maxViews: { type: integer }
        watermark: { type: boolean }
        allowDownload: { type: boolean }
    OfferDelegationRequest:
      type: object
      required: [grantorPartyId, granteePartyId, resourceType, resourceId, capabilities, grantScaSessionId]
      properties:
        grantorPartyId: { type: string, format: uuid }
        granteePartyId: { type: string, format: uuid }
        resourceType: { type: string, enum: [ACCOUNT, SAVINGS_GOAL, CARD, PAYMENT, STATEMENT, DOCUMENT] }
        resourceId: { type: string, format: uuid }
        capabilities:
          type: array
          items:
            type: string
            enum: [ACCOUNT_VIEW_DETAILS, ACCOUNT_READ_BALANCES, ACCOUNT_READ_TRANSACTIONS,
                   ACCOUNT_DOWNLOAD_STATEMENTS, ACCOUNT_INITIATE_PAYMENT, ACCOUNT_PROPOSE_PAYMENT,
                   ACCOUNT_MANAGE_BENEFICIARIES, ACCOUNT_MANAGE_LIMITS, CARD_VIEW,
                   CARD_VIEW_TRANSACTIONS, CARD_MANAGE_LIMITS, CARD_MANAGE_STATUS, CARD_MANAGE_CHANNELS, SAVINGS_DEPOSIT,
                   SAVINGS_WITHDRAW, SAVINGS_PROPOSE_WITHDRAW, OBJECT_READ, DELEGATION_MANAGE]
        approvalPolicy:
          type: string
          enum: [SOLO, ANY_ONE, ALL, N_OF_M]
          default: SOLO
          description: >-
            Only SOLO is enforced. Any other value is rejected with 400 APPROVAL_POLICY_UNSUPPORTED:
            no service counts approvals against a grant, so a co-signing requirement set here would
            be stored and echoed while a single owner decision still releases the money. The enum
            values remain listed because ADR-0232 D8 describes them as intended; they are not built.
          x-unsupported-values: [ANY_ONE, ALL, N_OF_M]
        requiredApprovals:
          type: integer
          deprecated: true
          description: >-
            Meaningful only with approvalPolicy=N_OF_M, which is rejected. Nothing counts approvals.
        perTransactionLimit:
          $ref: '#/components/schemas/MoneyDto'
          description: >-
            Per-amount ceiling. It never accumulates — it caps each payment and says nothing about
            how many there may be, which is why it does NOT satisfy the ADR-0249 D5 requirement
            that ACCOUNT_INITIATE_PAYMENT carry a cumulative ceiling.
        dailyLimit:
          $ref: '#/components/schemas/MoneyDto'
          description: |
            Cumulative ceiling per calendar day, counted authoritatively by the reservation API
            (ADR-0249 D3). Enforced from version 1.4.0; before that the field was rejected because
            nothing counted it.

            The day is a **Europe/Prague** calendar day, not a UTC one — a customer's "5 000 Kč per
            day" is their own day, and a UTC counter would roll over at 01:00/02:00 local and hand
            the delegate a second full ceiling.

            Rejected with 400 CUMULATIVE_LIMIT_UNSUPPORTED on a grant with no money-moving
            capability: nothing would ever reserve against it.
        monthlyLimit:
          $ref: '#/components/schemas/MoneyDto'
          description: >-
            As dailyLimit, over the Europe/Prague calendar month.
        exposure:
          $ref: '#/components/schemas/ExposureDto'
          deprecated: true
          description: >-
            Non-null values are refused with 400 EXPOSURE_UNSUPPORTED. Historical values remain
            readable only for audit and do not authorize access.
        validTo: { type: string, format: date-time }
        grantScaSessionId: { type: string, format: uuid }
        note: { type: string }
...
    DelegationLifecycleApprovalResponse:
      type: object
      required: [id, delegationId, operation, requestedReason, state, proposedBy, proposedAt]
      properties:
        id: { type: string, format: uuid }
        delegationId: { type: string, format: uuid }
        operation: { type: string, enum: [SUSPEND, REINSTATE, REVOKE] }
        requestedReason: { type: string }
        state:
          type: string
          enum: [PROPOSED, REJECTED, EXECUTED]
          description: EXECUTED means the revision-matched lifecycle transition and its outbox event committed atomically.
        proposedBy: { type: string }
        proposedAt: { type: string, format: date-time }
        expectedLifecycleRevision:
          type: [integer, 'null']
          format: int64
          minimum: 0
          description: Grant lifecycle revision inspected by the maker. Null legacy evidence is readable but cannot execute.
        decidedBy: { type: [string, 'null'] }
        decidedAt: { type: [string, 'null'], format: date-time }
        decisionReason: { type: [string, 'null'] }
        executedAt: { type: [string, 'null'], format: date-time }
...
    DelegationResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        grantorPartyId: { type: string, format: uuid }
        granteePartyId: { type: string, format: uuid }
        grantorName:
          type: string
          description: |
            Counterparty display name snapshotted onto the grant at offer time (issue #3604),
            from the pid-service eligibility lookup this service already performs. It is what was
            true when the grant was made, not a live value — an authorisation record must not be
            silently rewritten by a later rename.

            Null on grants offered before this field existed, and when pid-service returns no
            usable name. Clients fall back to the party id; they must not render a blank, which
            on a consent screen reads as a name that failed to load.

            Only ever returned to a party of the grant (the read paths are party-scoped) or to
            role-gated bank staff. This is not a party-name lookup.
        granteeName:
          type: string
          description: As grantorName, for the party receiving the grant.
        resourceType: { type: string }
        resourceId: { type: string, format: uuid }
        capabilities: { type: array, items: { type: string } }
        approvalPolicy: { type: string }
        requiredApprovals: { type: integer }
        perTransactionLimit: { $ref: '#/components/schemas/MoneyDto' }
        dailyLimit:
          $ref: '#/components/schemas/MoneyDto'
          deprecated: true
          description: >-
            Always null on grants offered after this API version — the field is refused on write.
            A non-null value belongs to a grant stored before that and is NOT enforced: it is a
            number on a record, not a cap on the delegate. Do not render it as a limit.
        monthlyLimit:
          $ref: '#/components/schemas/MoneyDto'
          deprecated: true
          description: >-
            See dailyLimit. Not enforced; do not render it as a limit.
        exposure:
          $ref: '#/components/schemas/ExposureDto'
          deprecated: true
          description: >-
            Historical audit metadata only. New non-null values are refused with 400
            EXPOSURE_UNSUPPORTED and never authorize access.
        validFrom: { type: string, format: date-time }
        validTo: { type: string, format: date-time }
        status:
          type: string
          enum: [OFFERED, ACTIVE, SUSPENDED, REVOKED, DECLINED, EXPIRED, RENOUNCED]
        note: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time }
        closedReason: { type: string }
...
    ReserveSpendRequest:
      type: object
      required: [amount, currency, idempotencyKey]
      properties:
        amount:
          type: string
          description: Decimal amount in major units, scaled to the currency's minor unit. Never a float.
        currency: { type: string, minLength: 3, maxLength: 3 }
        idempotencyKey:
          type: string
          maxLength: 200
          description: >-
            The caller's key for this spend, unique per delegation. Enforced by a database
            constraint, so a replay cannot double-count even across replicas.
        operationType:
          type: string
          enum: [UNSPECIFIED, DOMESTIC_PAYMENT]
          default: UNSPECIFIED
          description: Reconciliation rail; it is not workload authentication.
    SpendReservationResponse:
      type: object
      required: [reservationId, delegationId, amount, operationType, state, createdAt]
      properties:
        reservationId: { type: string, format: uuid }
        delegationId: { type: string, format: uuid }
        amount: { $ref: '#/components/schemas/MoneyDto' }
        operationType:
          type: string
          enum: [UNSPECIFIED, DOMESTIC_PAYMENT]
        state:
          type: string
          enum: [RESERVED, CONFIRMED, RELEASED]
          description: RESERVED and CONFIRMED consume headroom; RELEASED gives it back.
        createdAt:
          type: string
          format: date-time
          description: The window anchor — a reservation belongs to the day it took the headroom.
        settledAt: { type: string, format: date-time, nullable: true }
    SpendRefusalResponse:
      type: object
      required: [status, reason, error]
      properties:
        status: { type: integer }
        reason:
          type: string
          enum: [PER_TX, DAILY, MONTHLY, GRANT_NOT_ACTIVE, NO_SPEND_CAPABILITY, CURRENCY_MISMATCH]
        ceiling: { $ref: '#/components/schemas/MoneyDto' }
        alreadyCounted: { $ref: '#/components/schemas/MoneyDto' }
        remaining:
          $ref: '#/components/schemas/MoneyDto'
          description: What is still spendable under the breached ceiling, clamped at zero.
        error: { type: string }
    SpendIdempotencyConflictResponse:
      type: object
      required: [type, title, status, detail, code, error]
      properties:
        type: { type: string, enum: ['urn:openbank:error:idempotency-key-reused'] }
        title: { type: string, enum: [Idempotency key reused] }
        status: { type: integer, enum: [409] }
        detail: { type: string }
        code: { type: string, enum: [IDEMPOTENCY_KEY_REUSED] }
        error: { type: string }